🔒 [API] Authorize Todo
AOP for param
python server for nlp work

+++ java/src/main/java/app/beautyminder/controller/todo/TodoController.java
@@ -9,11 +9,14 @@ import app.beautyminder.dto.todo.TodoUpdateDTO;
 import app.beautyminder.service.MongoService;
 import app.beautyminder.service.TodoService;
 import app.beautyminder.service.auth.UserService;
+import app.beautyminder.util.AuthenticatedUser;
 import app.beautyminder.util.ValidUserId;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -23,26 +26,23 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/todo")
+@PreAuthorize("hasRole('ROLE_USER')")
 public class TodoController {
 
     private final TodoService todoService;
-    private final UserService userService;
     private final MongoService mongoService;
 
     @Operation(summary = "Retrieve all todos", description = "모든 Todo 항목을 검색합니다.", tags = {"Todo Operations"})
     @GetMapping("/all")
-    public Map<String, Object> getTodos(@ValidUserId String userId) {
-        User user = userService.findById(userId);
+    public Map<String, Object> getTodos(@AuthenticatedUser User user) {
         List<Todo> existingTodos = todoService.findTodosByUserId(user.getId());
         return createResponse("Here are the todos", existingTodos.isEmpty() ? Collections.emptyList() : existingTodos);
     }
 
     @Operation(summary = "Create a new todo", description = "새로운 Todo 항목을 추가합니다.", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Todo details for creation"), tags = {"Todo Operations"})
     @PostMapping("/create")
-    public ResponseEntity<AddTodoResponse> createTodo(@RequestBody AddTodoRequest request) {
+    public ResponseEntity<AddTodoResponse> createTodo(@RequestBody AddTodoRequest request, @AuthenticatedUser User user) {
         try {
-            var user = userService.findById(request.getUserId());
-
             // Check if a Todo already exists for this user with the same date
             if (todoService.existsByDateAndUserId(request.getDate(), user.getId())) {
                 return ResponseEntity.badRequest().body(new AddTodoResponse("Todo already exists for this date", null));
@@ -63,7 +63,9 @@ public class TodoController {
 
     @Operation(summary = "Update an existing todo by fields", description = "기존 Todo 항목을 업데이트합니다. (DB call 방식)", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Todo details for update"), tags = {"Todo Operations"})
     @PutMapping("/update/fields/{todoId}")
-    public ResponseEntity<AddTodoResponse> updateTodoByFields(@PathVariable String todoId, @RequestBody Map<String, Object> updates) {
+    public ResponseEntity<AddTodoResponse> updateTodoByFields(@PathVariable String todoId, @RequestBody Map<String, Object> updates, @AuthenticatedUser User user) {
+        todoService.checkUserAuthorizationForTodo(todoId, user.getId());
+
         Optional<Todo> updatedTodo = mongoService.updateFields(todoId, updates, Todo.class);
 
         return updatedTodo.map(todo -> ResponseEntity.ok(new AddTodoResponse("Updated Todo", todo))).orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).body(new AddTodoResponse("Failed to update Todo", null)));
@@ -71,7 +73,8 @@ public class TodoController {
 
     @Operation(summary = "Update an existing todo", description = "기존 Todo 항목을 업데이트합니다. (JSON 방식)", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Todo update"), tags = {"Todo Operations"})
     @PutMapping("/update/{todoId}")
-    public ResponseEntity<AddTodoResponse> updateTodoByTask(@PathVariable String todoId, @RequestBody TodoUpdateDTO todoUpdateDTO) {
+    public ResponseEntity<AddTodoResponse> updateTodoByTask(@PathVariable String todoId, @RequestBody TodoUpdateDTO todoUpdateDTO, @AuthenticatedUser User user) {
+        todoService.checkUserAuthorizationForTodo(todoId, user.getId());
         try {
             Optional<Todo> updatedTodo = todoService.updateTodoTasks(todoId, todoUpdateDTO);
 
@@ -82,14 +85,16 @@ public class TodoController {
     }
 
     @Operation(summary = "Delete a todo", description = "Todo 항목을 삭제합니다.", tags = {"Todo Operations"})
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deleteTodo(@PathVariable String id) {
+    @DeleteMapping("/delete/{todoId}")
+    public ResponseEntity<String> deleteTodo(@PathVariable String todoId, @AuthenticatedUser User user) {
+        todoService.checkUserAuthorizationForTodo(todoId, user.getId());
+
         try {
-            boolean deleted = todoService.deleteTodoById(id);
+            boolean deleted = todoService.deleteTodoById(todoId);
             if (deleted) {
                 return ResponseEntity.ok("Todo deleted successfully");
             } else {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Todo not found with id: " + id);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Todo not found with id: " + todoId);
             }
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
@@ -98,7 +103,9 @@ public class TodoController {
 
     @Operation(summary = "Delete a task", description = "Task를 삭제합니다. (use /delete/id)", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Todo update"), tags = {"Todo Operations"})
     @DeleteMapping("/delete/{todoId}/task/{taskId}")
-    public ResponseEntity<String> deleteTask(@PathVariable String todoId, @PathVariable String taskId) {
+    public ResponseEntity<String> deleteTask(@PathVariable String todoId, @PathVariable String taskId, @AuthenticatedUser User user) {
+        todoService.checkUserAuthorizationForTodo(todoId, user.getId());
+
         try {
             boolean taskDeleted = todoService.deleteTaskFromTodoById(todoId, taskId);
             if (taskDeleted) {
